Fixed banner so it appears on Accounts Settings view
